+++ ChangeLog.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 
 - functions `get-request-parameter`, and `get-request-long-parameter` to 
   :ring-util module
+- more HTTP status codes
   
 ### Improved 
 
@@ -21,7 +22,7 @@ All notable changes to this project will be documented in this file.
 
 ### Fixed 
 
-- HTTP client SSL requests, added more status codes 
+- HTTP client SSL requests